For 2.7.0 Library
For used for ESP32 DEVKIT V1 and use the same old board. This library is the best suitable for that. Use Arduino IDE for best outcome. Incase of PlatformIO prefer mentioning the right version to be included in your project. (This library does not include many other functions. So the more you advance, you will consider fixing the problem and then upgrading your library.)